+++ BackEnd/backend/src/main/java/com/example/backend/controller/MultimediaContentController.java
@@ -21,14 +21,14 @@ public class MultimediaContentController {
     private MultimediaContentService multimediaContentService;
 
     @PostMapping("/upload")
-    public ResponseEntity<MultimediaContentResponse> uploadFile(@RequestBody MultimediaContentRequest request, @RequestParam int idPoi) {
+    public ResponseEntity<MultimediaContentResponse> uploadFile(@RequestBody MultimediaContentRequest request, @RequestParam Long idPoi) {
         MultimediaContent multimediaContent = multimediaContentService.saveMultimediaContent(request.toMultimediaContent(), request.authorFirstName(), request.authorLastName(), request.mediaFile(), idPoi);
         MultimediaContentResponse response = MultimediaContentResponse.mapToResponse(multimediaContent);
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
 
     @GetMapping("/find/{id}")
-    public ResponseEntity<MultimediaContentResponse> getContentDetails(@PathVariable int id) {
+    public ResponseEntity<MultimediaContentResponse> getContentDetails(@PathVariable Long id) {
         MultimediaContent multimediaContent = multimediaContentService.getMultimediaContentById(id);
         MultimediaContentResponse response = MultimediaContentResponse.mapToResponse(multimediaContent);
         return ResponseEntity.status(HttpStatus.OK).body(response);
@@ -42,14 +42,14 @@ public class MultimediaContentController {
     }
 
     @PutMapping("/edit/content/{idContent}")
-    public ResponseEntity<MultimediaContentResponse> updateContent(@PathVariable int idContent, @RequestBody MultimediaContent multimediaContentDetails) {
+    public ResponseEntity<MultimediaContentResponse> updateContent(@PathVariable Long idContent, @RequestBody MultimediaContent multimediaContentDetails) {
         MultimediaContent multimediaContent = multimediaContentService.updateContent(idContent, multimediaContentDetails);
         MultimediaContentResponse response = MultimediaContentResponse.mapToResponse(multimediaContent);
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
 
     @PutMapping("/edit/file/{idFile}")
-    public ResponseEntity<MultimediaContentResponse> updateFile(@PathVariable int idFile, @RequestParam int idContent,
+    public ResponseEntity<MultimediaContentResponse> updateFile(@PathVariable Long idFile, @RequestParam Long idContent,
                                                                 @RequestBody MediaFile mediaFile) {
 
         MultimediaContent multimediaContent = multimediaContentService.updateFile(idContent, idFile, mediaFile);
@@ -58,7 +58,7 @@ public class MultimediaContentController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Void> deleteMultimediaContent(@PathVariable int id)  {
+    public ResponseEntity<Void> deleteMultimediaContent(@PathVariable Long id)  {
         multimediaContentService.deleteMultimediaContentById(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
@@ -77,7 +77,7 @@ public class MultimediaContentController {
     }
 
     @PatchMapping("/validate/{id}")
-    public ResponseEntity<MultimediaContentResponse> validateMultimediaContent(@PathVariable int id, @RequestParam Status status) {
+    public ResponseEntity<MultimediaContentResponse> validateMultimediaContent(@PathVariable Long id, @RequestParam Status status) {
         MultimediaContent multimediaContent = multimediaContentService.validateMultimediaContent(id, status);
         MultimediaContentResponse response = MultimediaContentResponse.mapToResponse(multimediaContent);
         return ResponseEntity.status(HttpStatus.OK).body(response);
