+++ BackEnd/backend/src/main/java/com/example/backend/repository/ContestRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ContestRepository extends JpaRepository<Contest, Integer> {
+public interface ContestRepository extends JpaRepository<Contest, Long> {
 
     List<Contest> searchByName(String name);
