+++ BackEnd/backend/src/main/java/com/example/backend/repository/MediaFileRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MediaFileRepository extends JpaRepository<MediaFile, Integer> {
+public interface MediaFileRepository extends JpaRepository<MediaFile, Long> {
 }
