+++ BackEnd/backend/src/main/java/com/example/backend/repository/PointOfInterestRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface PointOfInterestRepository extends JpaRepository<PointOfInterest, Integer> {
+public interface PointOfInterestRepository extends JpaRepository<PointOfInterest, Long> {
 
     Optional<PointOfInterest> findByNameAndLatitudeAndLongitude(String name, double lat, double lon);
     List<PointOfInterest> searchByName(String name);
